old-tests/submission/Opera/preload/metadata/preload-metadata-to-none-after-loadstart-readystate-mp4.html

<!doctype html>
<!-- This file was generated from 'preload-metadata-to-none-after-loadstart-readystate.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:metadata readyState when preload is set to none after loadstart is fired - mp4</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata" controls></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:metadata readyState when preload is set to none after loadstart is fired - mp4", "HTMLVideoElement.preload:metadata readyState when preload is set to none after loadstart is fired - mp4", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
	var tag_name = vars[1];
	var states_actual = '';
	var node = document.getElementsByTagName(tag_name)[0];
	node.addEventListener('loadstart', function(){ node.preload='none'; }, false);
	node.addEventListener('suspend', endTest, false);
	node.addEventListener('error', function() {
	    if (!node.canPlayType(tag_name + '/mp4')) {
		t.step(function() {
		    assert_unreached("Cannot play '" + tag_name + "/mp4'");
		    t.done();
		});
	    }
	}, false);
	node.src = '../resources/preload.mp4?pipe=trickle(100000:d1:r2)&nocache=' + Math.random();

	function endTest() {
	    t.step(function() {
		assert_equals(node.readyState, HTMLMediaElement.HAVE_CURRENT_DATA, 'readyState after suspend')
		t.done();
	    });
	}
    });
});
</script>
